# packages/service-v2/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/health": {
      "get": {
        "operationId": "HealthController_getHealth",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Health check endpoint",
        "tags": [
          "health"
        ]
      }
    },
    "/api/units": {
      "get": {
        "operationId": "UnitsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List units",
        "tags": [
          "units"
        ]
      },
      "post": {
        "operationId": "UnitsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUnitDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a unit",
        "tags": [
          "units"
        ]
      }
    },
    "/api/units/{id}": {
      "get": {
        "operationId": "UnitsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a unit by ID",
        "tags": [
          "units"
        ]
      },
      "patch": {
        "operationId": "UnitsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUnitDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a unit",
        "tags": [
          "units"
        ]
      },
      "delete": {
        "operationId": "UnitsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Delete a unit",
        "tags": [
          "units"
        ]
      }
    },
    "/api/meter-types": {
      "get": {
        "operationId": "MeterTypesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List meter types",
        "tags": [
          "meter-types"
        ]
      },
      "post": {
        "operationId": "MeterTypesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMeterTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a meter type",
        "tags": [
          "meter-types"
        ]
      }
    },
    "/api/meter-types/{id}": {
      "get": {
        "operationId": "MeterTypesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a meter type by ID",
        "tags": [
          "meter-types"
        ]
      },
      "patch": {
        "operationId": "MeterTypesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeterTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a meter type",
        "tags": [
          "meter-types"
        ]
      },
      "delete": {
        "operationId": "MeterTypesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Delete a meter type",
        "tags": [
          "meter-types"
        ]
      }
    },
    "/api/meters": {
      "get": {
        "operationId": "MetersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List meters",
        "tags": [
          "meters"
        ]
      },
      "post": {
        "operationId": "MetersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMeterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a meter",
        "tags": [
          "meters"
        ]
      }
    },
    "/api/meters/{id}": {
      "get": {
        "operationId": "MetersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a meter by ID",
        "tags": [
          "meters"
        ]
      },
      "patch": {
        "operationId": "MetersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a meter",
        "tags": [
          "meters"
        ]
      },
      "delete": {
        "operationId": "MetersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Delete a meter",
        "tags": [
          "meters"
        ]
      }
    },
    "/api/values": {
      "get": {
        "operationId": "ValuesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List values (latest first)",
        "tags": [
          "values"
        ]
      },
      "post": {
        "operationId": "ValuesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateValueDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a value",
        "tags": [
          "values"
        ]
      }
    },
    "/api/values/{id}": {
      "get": {
        "operationId": "ValuesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a value by ID",
        "tags": [
          "values"
        ]
      },
      "patch": {
        "operationId": "ValuesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateValueDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a value",
        "tags": [
          "values"
        ]
      },
      "delete": {
        "operationId": "ValuesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Delete a value",
        "tags": [
          "values"
        ]
      }
    },
    "/api/notes": {
      "get": {
        "operationId": "NotesController_findAll",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "ISO timestamp lower bound",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "ISO timestamp upper bound",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List notes (supports optional from/to timestamp filtering)",
        "tags": [
          "notes"
        ]
      },
      "post": {
        "operationId": "NotesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a note",
        "tags": [
          "notes"
        ]
      }
    },
    "/api/notes/{id}": {
      "get": {
        "operationId": "NotesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a note by ID",
        "tags": [
          "notes"
        ]
      },
      "patch": {
        "operationId": "NotesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNoteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a note",
        "tags": [
          "notes"
        ]
      },
      "delete": {
        "operationId": "NotesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Delete a note",
        "tags": [
          "notes"
        ]
      }
    },
    "/api/reports/{meterTypeId}": {
      "get": {
        "operationId": "ReportsController_findByMeterType",
        "parameters": [
          {
            "name": "meterTypeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "ISO timestamp lower bound",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "ISO timestamp upper bound",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get value report rows for a meter type (from v_value_stats)",
        "tags": [
          "reports"
        ]
      }
    }
  },
  "info": {
    "title": "Meter App Service V2",
    "description": "REST API for meter app",
    "version": "0.1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateUnitDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Kilowattstunde"
          },
          "suffix": {
            "type": "string",
            "example": "kWh"
          }
        },
        "required": [
          "name",
          "suffix"
        ]
      },
      "UpdateUnitDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Kilowattstunde"
          },
          "suffix": {
            "type": "string",
            "example": "kWh"
          }
        }
      },
      "CreateMeterTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Strom"
          },
          "unitId": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "name",
          "unitId"
        ]
      },
      "UpdateMeterTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Strom"
          },
          "unitId": {
            "type": "number",
            "example": 1
          }
        }
      },
      "CreateMeterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Main electricity meter"
          },
          "meterTypeId": {
            "type": "number",
            "example": 2
          }
        },
        "required": [
          "name",
          "meterTypeId"
        ]
      },
      "UpdateMeterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Main electricity meter"
          },
          "meterTypeId": {
            "type": "number",
            "example": 2
          }
        }
      },
      "CreateValueDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "number",
            "example": 1245.67
          },
          "timestamp": {
            "type": "string",
            "example": "2026-03-05T12:00:00.000Z"
          },
          "meterId": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "value",
          "timestamp",
          "meterId"
        ]
      },
      "UpdateValueDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "number",
            "example": 1245.67
          },
          "timestamp": {
            "type": "string",
            "example": "2026-03-05T12:00:00.000Z"
          },
          "meterId": {
            "type": "number",
            "example": 1
          }
        }
      },
      "CreateNoteDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "example": "Yearly meter check done"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-03-05T12:00:00.000Z"
          }
        },
        "required": [
          "text",
          "timestamp"
        ]
      },
      "UpdateNoteDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "example": "Yearly meter check done"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-03-05T12:00:00.000Z"
          }
        }
      }
    }
  }
}