# packages/service-v2/src/notes/dto/create-note.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsDateString, IsNotEmpty, IsString, MaxLength } from 'class-validator';

export class CreateNoteDto {
  @ApiProperty({ example: 'Yearly meter check done' })
  @IsString()
  @IsNotEmpty()
  @MaxLength(2000)
  text!: string;

  @ApiProperty({ example: '2026-03-05T12:00:00.000Z' })
  @IsDateString()
  timestamp!: string;
}